# api/internal/docs/generated/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TheTipTop API",
        "title": "TheTipTop",
        "contact": {},
        "version": "dev"
    },
    "host": "localhost",
    "paths": {
...
        "/user/auth": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Authenticate a client/employees.",
                "operationId": "user.UserAuth",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "default": "user-thetiptop@yopmail.com",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Aa1@azetyuiop",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client signed in"
                    },
                    "400": {
                        "description": "Invalid email or password"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/auth/renew": {
            "get": {
                "consumes": [
                    "*/*",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Renew JWT for a client/employees.",
                "operationId": "user.UserAuthRenew",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token renewed"
                    },
                    "400": {
                        "description": "Invalid token"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update a client/employees password.",
                "operationId": "jwt.Auth =\u003e user.CredentialUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "default": "user-thetiptop@yopmail.com",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Aa1@azetyuiop",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Password updated"
                    },
                    "400": {
                        "description": "Invalid email, password or token"
                    },
                    "404": {
                        "description": "Client not found"
                    },
                    "409": {
                        "description": "Client already validated"
                    },
                    "410": {
                        "description": "Token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/register/validation": {
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Validate a client/employees email.",
                "operationId": "user.MailValidation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "default": "user-thetiptop@yopmail.com",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client email validate"
                    },
                    "400": {
                        "description": "Invalid email or token"
                    },
                    "404": {
                        "description": "Client not found"
                    },
                    "409": {
                        "description": "Client already validated"
                    },
                    "410": {
                        "description": "Token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/validation/renew": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Recover a client/employees validation type.",
                "operationId": "user.ValidationRecover",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "default": "user-thetiptop@yopmail.com",
                        "description": "Email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "mail",
                            "password",
                            "phone"
                        ],
                        "type": "string",
                        "description": "Type of validation",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}